Small tweaks to projectiles

+++ Assets/Scripts/Items/Projectile Types/TestParticle.cs
@@ -8,7 +8,7 @@ public class TestParticle : BaseProjectile
 	public ParticleSystem contact;
 
 	// Use this for initialization
-	public override void Setup (GameObject shooter, float damageMod = 1.0f, float speedMod = 1.0f, int rotationMod = 0) 
+	public override void Setup (GameObject shooter, float damageMod = 1.0f, float speedMod = 1.0f, float rotationMod = 0) 
     {
 	    speed = 5; 
         lifespan = 3;
